+++ app/controllers/books_controller.rb
@@ -56,7 +56,7 @@ class BooksController < ApplicationController
   private
 
  def book_params
-   params.require(:book).permit(:title, :body)
+   params.require(:book).permit(:title, :body, :rate)
  end
 
   def ensure_correct_user
@@ -65,6 +65,7 @@ class BooksController < ApplicationController
      redirect_to "/books"
     end
   end
+  
   
   
 end
