+++ app/views/books/_form.html.erb
@@ -11,7 +11,7 @@
         </div>
         
   <div class="form-group row" id="star">
-    <%= f.label :rate,'評価 ', class:'col-md-3 col-form-label' %>
+    <%= f.label :rate, class:'col-md-3 col-form-label' %>
     <%= f.hidden_field :rate, id: :review_star %>
   </div>
