+++ app/views/books/edit.html.erb
@@ -26,16 +26,19 @@
   </div>
 
   <!-- 評価javascript -->
-  <script>
-  $('#star').raty({
-    size     : 36,
-    starOff:  '<%= asset_path('star-off.png') %>',
-    starOn : '<%= asset_path('star-on.png') %>',
-    starHalf: '<%= asset_path('star-half.png') %>',
-    scoreName: 'book[rate]',
-    half: true,
-  });
-  </script>  
+          <div id="rate_<%= @book.id %>"></div>
+          <script>
+          $('#rate_<%= @book.id %>').raty({
+            size: 20,
+            starOff:  '<%= asset_path('star-off.png') %>',
+            starOn : '<%= asset_path('star-on.png') %>',
+            starHalf: '<%= asset_path('star-half.png') %>',
+            half: true,
+            readOnly: true,
+            score: <%= @book.rate %> , 
+            });
+          </script>
+
 
 
  <%= link_to "Show", book_path(@book.id) %>|<%= link_to "Back", '/books' %>
